Cast operator/function arguments into expected types of the chosen form
The callable form was chosen via an implicit cast, cast the arguments
so that the backend has no wiggle room to apply its own (potentially
different) casting.

Fixes: #355.
